[Chore #163533980] Prepare the project's environment configuration and structure
