--- Procfile
@@ -1,1 +1,0 @@
-web: DEBUG=readit:start,readit:db NODE_ENV=production nodemon --exec babel-node ./server/index.js
