--- config/config.js
@@ -1,23 +1,0 @@
-export default {
-  development: {
-    username: 'musigwa',
-    password: 'pro@247DB',
-    database: 'sendit',
-    host: '127.0.0.1',
-    dialect: 'postgres',
-  },
-  test: {
-    username: 'musigwa',
-    password: 'pro@247DB',
-    database: 'sendit',
-    host: '127.0.0.1',
-    dialect: 'postgres',
-  },
-  production: {
-    username: 'musigwa',
-    password: 'pro@247DB',
-    database: 'sendit',
-    host: '127.0.0.1',
-    dialect: 'postgres',
-  },
-};
