--- config/custom-environment-variables.json
@@ -1,6 +1,0 @@
-{
-  "PORT": "PORT",
-  "JWT_KEY": "READIT_JWT",
-  "DATABASE_URL": "DATABASE_URL",
-  "DEBUGGER": "DEBUG"
-}
