--- config/default.json
@@ -1,3 +1,0 @@
-{
-  "PORT": 5070
-}
