--- config/development.json
@@ -1,4 +1,0 @@
-{
-  "DATABASE_URL": "postgresql://musigwa:pro@247DB@localhost:5432/sendit",
-  "JWT_KEY": "SET_YOUR_KEY"
-}
