+++ package.json
@@ -1,15 +1,14 @@
 {
-<<<<<<< HEAD
   "name": "ReadIT",
   "version": "1.0.0",
   "description": "Endpoint API for ReadIT application",
-  "main": "./server/index.js",
+  "main": "index.js",
   "author": "Titan-Devs",
   "license": "MIT",
   "scripts": {
     "testenv": "NODE_ENV=test",
     "devenv": "NODE_ENV=development",
-    "server": "clear && DEBUG=readit:* yarn devenv nodemon --exec babel-node ./server/index.js",
+    "server": "clear && DEBUG=readit:* yarn devenv nodemon --exec babel-node index.js",
     "build": "babel server -d server/dist",
     "coveralls": "jest --coverage && cat ./tests/coverage/lcov.info | coveralls",
     "test": "clear && yarn testenv jest --coverage"
@@ -26,9 +25,9 @@
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "bcrypt": "^3.0.2",
-    "config": "^3.0.0",
     "cross-env": "5.2.0",
     "debug": "^4.1.0",
+    "dotenv": "^6.2.0",
     "express": "4.16.4",
     "joi": "^14.0.4",
     "jsonwebtoken": "8.3.0",
@@ -51,33 +50,4 @@
     "prettier-eslint": "^8.8.2",
     "prettier-eslint-cli": "^4.7.1"
   }
-=======
-  "name": "readit",
-  "version": "1.0.0",
-  "main": "index.js",
-  "directories": {
-    "test": "tests"
-  },
-  "dependencies": {
-    "dotenv": "^6.2.0",
-    "express": "^4.16.4",
-    "pg": "^7.8.0",
-    "sequelize": "^4.42.0"
-  },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/abayo-luc/readIt.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/abayo-luc/readIt/issues"
-  },
-  "homepage": "https://github.com/abayo-luc/readIt#readme",
-  "description": ""
->>>>>>> a3e64babe2a27bb5e74b169300bd42490f493849
 }
