--- server/models/User.js
@@ -1,20 +1,0 @@
-import Sequelize from 'sequelize';
-import Database from '../database';
-
-const User = Database.connection().define(
-  'imigati',
-  {
-    firstName: {
-      type: Sequelize.STRING,
-    },
-    lastName: {
-      type: Sequelize.STRING,
-      allowNull: false,
-    },
-  },
-  { timestamps: true },
-);
-
-// force: true will drop the table if it already exists
-// Table created
-User.sync({ force: true }).then(() => {});
