--- server/models/index.js
@@ -1,33 +1,0 @@
-import fs from 'fs';
-import path from 'path';
-import Sequelize from 'sequelize';
-import configs from '../../config/config';
-
-const basename = path.basename(__filename);
-const db = {};
-const env = process.env.NODE_ENV || 'development';
-const config = configs[env];
-
-const sequelize = process.env.DATABASE_URL
-  ? new Sequelize(process.env.DATABASE_URL, config)
-  : new Sequelize(config.database, config.username, config.password, config);
-
-fs.readdirSync(__dirname)
-  .filter(
-    file => file.indexOf('.') !== 0 && file !== basename && file.slice(-3) === '.js',
-  )
-  .forEach((file) => {
-    const model = sequelize.import(path.join(__dirname, file));
-    db[model.name] = model;
-  });
-
-Object.keys(db).forEach((modelName) => {
-  if (db[modelName].associate) {
-    db[modelName].associate(db);
-  }
-});
-
-db.sequelize = sequelize;
-db.Sequelize = Sequelize;
-
-export default db;
