--- server/tests/testData.js
@@ -1,5 +1,0 @@
-const newUser = {};
-
-const baseUrl = '/api/v1';
-
-export { newUser, baseUrl };
